Trace sampler interface is now just a Proc
